ZCO-SmartPhone.cpp

+++ ZCO-SmartPhone.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
 	int n;
 	cin >> n;
-	int a[n],b[n];
+	long a[n],b[n];
 	for(int i =0 ; i<n;i++){
 	    cin >> a[i];
 	}
